app/requirements.txt
streamlit
pillow
websocket-client
threading